Changed routing table array to static data type. Added method to retrieve routing table entries. Changed project packet payload of UPDATE packet to type short. Untested method for updating an entry in the routing table class. Test packet generator for making new payload for UPDATE packet.

+++ new_routing_table_method/elements/proj_packets.hh
@@ -9,7 +9,7 @@ struct PacketUPDATE{
 	 uint16_t source;
 	 uint8_t sequence;
 	 uint16_t length;
-     uint64_t payload;
+     unsigned short payload;
 };
 
 struct PacketACK{
